requirements.txt
streamlit
PyPDF2
numpy
scikit-learn
google-generativeai
requests
pytesseract
pdf2image
Pillow
pypdfium2
transformers
torch
python-dotenv